% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompPrinciple.R
\name{comp.principle}
\alias{comp.principle}
\title{Compounding Principle}
\usage{
comp.principle(A, N, T)
}
\arguments{
\item{A=Priciple, }{N= Interest Rate YOY in %, T= Number of Years}
}
\value{
Final Principle
}
\description{
Compounding Principle
}
\author{
Nikhil Siddavanahalli Prakash
}
